2.Proyectos/Proyecto_1Contacto/requirements.txt
blinker==1.9.0
certifi==2025.4.26
click==8.2.0
cloudinary==1.44.0
colorama==0.4.6
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
